showing self transfers as such
