+++ src/components/TransferTable.vue
@@ -4,6 +4,7 @@ import DateField from 'components/DateField';
 import TransactionField from 'components/TransactionField';
 import { formatWei, toChecksumAddress } from 'src/lib/utils';
 import { getIcon } from 'src/lib/token-utils';
+import { getDirection } from 'src/lib/transaction-utils';
 
 export default {
     name: 'TransferTable',
@@ -89,6 +90,7 @@ export default {
             showDateAge: true,
             tokenList: {},
             highlightAddress: '',
+            getDirection,
         };
     },
     created(){
@@ -268,14 +270,10 @@ export default {
                 <DateField :epoch="convertToEpoch(props.row.timestamp)" :force-show-age="showDateAge"/>
             </q-td>
             <q-td key="direction" :props="props">
-                <span v-if="toChecksumAddress(address) === toChecksumAddress(props.row.from)" class="direction out">
-                    {{ $t('components.transaction.out').toUpperCase() }}
-                </span>
                 <span
-                    v-else-if="toChecksumAddress(address) === toChecksumAddress(props.row.to)"
-                    class="direction in"
+                    :class="`direction ${getDirection(address, props.row)}`"
                 >
-                    {{ $t('components.transaction.in').toUpperCase() }}
+                    {{ $t(`components.transaction.${getDirection(address, props.row)}`).toUpperCase() }}
                 </span>
             </q-td>
             <q-td key="from" :props="props">
@@ -400,22 +398,7 @@ export default {
 
 <style lang='scss' scoped>
 .direction {
-  user-select: none;
-  padding: 3px 6px;
-  border-radius: 5px;
-  font-size: 0.9em;
-
-  &.in {
-    color: rgb(0, 161, 134);
-    background: rgba(0, 161, 134, 0.1);
-    border: 1px solid rgb(0, 161, 134);
-  }
-
-  &.out {
-    color: #cc9a06 !important;
-    background: rgba(255, 193, 7, 0.1);
-    border: 1px solid #cc9a06 !important;
-  }
+  @include direction;
 }
 
 .nft-icon {
