+++ src/i18n/de-de/index.js
@@ -271,6 +271,7 @@ export default {
         transaction: {
             in: 'in',
             out: 'out',
+            self: 'self',
             load_error: 'Transaktionen konnten nicht geladen werden',
             show_short: 'Kurz anzeigen',
             show_total: 'Summe anzeigen',
