+++ src/i18n/en-us/index.js
@@ -275,6 +275,7 @@ export default {
         transaction: {
             in: 'in',
             out: 'out',
+            self: 'self',
             load_error: 'Could not load transactions',
             show_short: 'Show short',
             show_total: 'Show total',
