+++ src/i18n/es-es/index.js
@@ -272,6 +272,7 @@ export default {
         transaction: {
             in: 'entra',
             out: 'sale',
+            self: 'neutral',
             load_error: 'No se pudieron cargar las transacciones',
             show_short: 'Mostrar corta',
             show_total: 'Mostrar total',
