+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from 'src/types/BlockData';
 export * from 'src/types/EvmTransactionExtended';
 export * from 'src/types/LatestContainerOptions';
 export * from 'src/types/Pagination';
+export * from 'src/types/NftTransfers';
